refactoring link client and server
