+++ LoUAM/LoUAM.csproj
@@ -122,7 +122,7 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Link\LinkClient.cs" />
+    <Compile Include="Link\Link.cs" />
     <Compile Include="ControlPanel\EditPlace.xaml.cs">
       <DependentUpon>EditPlace.xaml</DependentUpon>
     </Compile>
@@ -135,7 +135,6 @@
     </Compile>
     <Compile Include="Map\MapImage.cs" />
     <Compile Include="Map\MapPlace.cs" />
-    <Compile Include="Link\LinkServer.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
